requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-filter==24.3